pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "game-art-pipeline"
version = "0.1.0"
description = "AI-powered game art generation — sprites, spritesheets, tilesets, and more"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]
diffusers = ["torch", "diffusers", "transformers", "accelerate"]

[project.scripts]
game-art = "game_art.cli:main"

[tool.setuptools.packages.find]
include = ["game_art*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
